requirements.txt
discord.py>=1.3.0
bs4>=0.0.1
gensim>=3.8.1
python-dateutil>=2.8.1
